refactor(registry): isolate legacy postadd execution
